janasamparka/backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Environment and configuration
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# HTTP client for external APIs
httpx==0.25.1
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1  # Parallel testing
pytest-mock==3.12.0  # Mocking support
pytest-timeout==2.2.0  # Timeout for slow tests
httpx==0.25.1  # Async HTTP client for testing

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Firebase Admin SDK (for auth and push notifications)
firebase-admin==6.3.0

# AI/ML
openai==1.3.7
sentence-transformers==2.2.2
numpy<2.0.0  # Lock to NumPy 1.x for compatibility

# Additional utilities
pytz==2023.3
python-dateutil==2.8.2
pillow==10.1.0
aiofiles==23.2.1
shapely==2.0.2
geopy==2.4.1
faiss-cpu==1.7.4  # For AI duplicate detection (Phase 2.5)

# Monitoring & Performance
psutil==5.9.6  # System monitoring
sentry-sdk==1.38.0  # Error tracking
structlog==23.2.0  # Structured logging
redis==5.0.1  # Caching and sessions
locust==2.17.0  # Load testing
python-json-logger==4.0.0  # JSON logging support
prometheus-client==0.23.1  # Prometheus metrics
opentelemetry-api==1.38.0  # OpenTelemetry API
opentelemetry-sdk==1.38.0  # OpenTelemetry SDK
